introduce IFF_NO_UNICAST_FLT
Patchwork-id: 65443
O-Subject: [Patch RHEL6 50/50 V3] net: introduce IFF_NO_UNICAST_FLT
Bugzilla: 791284 865301 950918 983117
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Jiri Benc <jbenc@redhat.com>

Upstream introduced the following changeset:

commit 01789349ee52e4a3faf376f1485303d9723c4f1f
Author: Jiri Pirko <jpirko@redhat.com>
Date:   Tue Aug 16 06:29:00 2011 +0000

    net: introduce IFF_UNICAST_FLT private flag

    Use IFF_UNICAST_FTL to find out if driver handles unicast
    address filtering. In case it does not, promisc mode is entered.

    Patch also fixes following drivers:
    stmmac, niu: support uc filtering and yet it propagated
            ndo_set_multicast_list
    bna, benet, pxa168_eth, ks8851, ks8851_mll, ksz884x : has set
	    ndo_set_rx_mode but do not support uc filtering

    Signed-off-by: Jiri Pirko <jpirko@redhat.com>
    Signed-off-by: David S.  Miller <davem@davemloft.net>

So, the mlx4 driver update has need for this infrastructure.  Of
course, we can't just backport the upstream fix because it would
break out of box drivers (namely, no out of box driver will have
IFF_UNICAST_FLT set, instead they rely on set_rx_mode being
defined in order to signal to the core network layer that they
can do unicast filtering).  So reverse the logic of the flag.
This way, instead of having to set the flag before the core
net code decides whether or not to see if we need to enter
promiscuous mode and patch up all the drivers that support, only
a driver that has been patched up and specifically wants to signal
to the core net code not to use the rx_set_mode handler in leiu
of promiscuous mode will set the flag.  This also reduces the
overall size of the patch considerably.

In the original patch, all of the drivers that would normally have
set IFF_UNICAST_FLT get ignored in this patch as the fact that
they define ndo_set_rx_mode makes them work as they should with
this reversed logic.

I fixed up the stmmac driver.  It didn't apply totally cleanly,
but I confirmed that the gmac supports unicast filtering, and
so on gmac devices the flag is unset, while on mac100 devices it
will get set.  This was the only driver I could find that didn't
want the same setting universally across all devices it supported.

The original patch lists a number of drivers that it says set
ndo_set_rx_mode but does not support uc filtering.  I can't speak
to all of them, but I did check the benet driver, and it sets
ndo_set_rx_mode and *does* support uc filtering.  In the name
of not destabalizing rhel, I did not change any of those drivers.
Either they work now, or they don't, or we don't even ship them.

This allows us to have a roughly equivalent infrastructure in
rhel, but without breaking kABI.  Driver updates that need this
infrastructure will simply need to reverse their usage of the
flag from how they would use it in the upstream driver in order
for it to work on rhel.

Version 2:

Oops, forgot to use netdev_extended(dev)->ext_priv_flags
Clean up the comment around the flag in netdevice.h

Version 3:

Remove the niu.c changes

Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>