lpfc: Merge with "lpfc 8.3.31: Fixed system panic due to midlayer abort and
Patchwork-id: 65554
O-Subject: [RHEL6.5 PATCH 83/84] scsi: lpfc: Merge with "lpfc 8.3.31: Fixed system panic due to midlayer abort and
Bugzilla: 947546
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Rob Evers on behalf of Emulex <revers@redhat.com>

driver complete race on SCSI cmd"

  This patch is specify to the branch to bring the release in sync with the upstream
code.  When the patch for "[SCSI] lpfc 8.3.31: Fixed system panic due to midlayer abort and
driver complete race on SCSI cmd" was submitted a spacing issue in the lpfc_crtn.h was
corrected in the patch.  However, the code released to redhat contained a extra space.
This patch corrects the spacing so the code is in sync with the upstream code.

Signed-off-by: Rafael Aquini <aquini@redhat.com>