mpt3sas: Added a driver module parameter max_msix_vectors
Patchwork-id: 65604
O-Subject: [RHEL6.5 e-stor PATCH] mpt3sas: Added a driver module parameter max_msix_vectors
Bugzilla: 997045
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Patch is posted to scsi-misc by LSI here -
http://marc.info/?l=linux-scsi&m=137648480415650&w=2
The change to the ver string is RH specific, obviously.

Tomas

Added a driver module parameter max_msix_vectors. Using this module parameter
the maximum number of MSI-X vectors could be set.

The number of MSI-X vectors used would be the minimum of MSI-X vectors supported by
the HBA, the number of CPU cores and the value set to max_msix_vectors module parameter.

The default value of this module parameter is set to 8. The default value of this parameter
is set to 8 inorder to reduce the amount of memory required for Reply Descriptor Post queue.
This is because with the higher MSI-X vectors, some times kernel is not able to allocate the
requested amount of memory and crash is observed. To overcome this problem,
the default value is set to 8.

Signed-off-by: Sreekanth Reddy <Sreekanth.Reddy@lsi.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>