kernel: z90crypt module load crash
Patchwork-id: 65606
O-Subject: [RHEL6.5 PATCH] [s390x] kernel: z90crypt module load crash
Bugzilla: 997621
RH-Acked-by: Steve Best <sbest@redhat.com>

Description
-----------
The kernel crashes with an oops error message after the
z90crypt module has been loaded.

The module reference counting of the sub-modules for the
type 6 and type 50 message operations uses a dummy module
structure if the zcrypt driver is built as a single,
monolithic module. The dummy module structures are not
initialized correctly.

Remove the dummy module structures and add the module
name to the message type operations structure instead.

Bugzilla
--------
BZ 997621
https://bugzilla.redhat.com/show_bug.cgi?id=997621

Upstream status of the patch
----------------------------
The patch is specific to RHEL6 and, hence, is not applicable
to the upstream kernel.

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Rafael Aquini <aquini@redhat.com>