dm-snapshot: fix data corruption
Patchwork-id: 5111
O-Subject: [kernel team] EMBARGOED RHEL6 bz995071: dm snapshot: fix data corruption
Bugzilla: 974481
CVE: CVE-2013-4299
RH-Acked-by: Alasdair G Kergon <agk@redhat.com>
RH-Acked-by: Petr  Matousek <pmatouse@redhat.com>

dm snapshot: fix data corruption

When we allocate a new chunk in persistent_prepare, we increment
ps->next_free and we make sure that it doesn't point to a metadata area.
If it does, it is further incremented.

When we load metadata from disk on device activation, ps->next_free is
positioned after the last used data chunk. However, if this last used data
chunk is followed by a metadata area, ps->next_free is positioned
errorneously to the metadata area. A newly allocated chunk is placed at
the same location as the metadata area, this results in data or metadata
corruption.

This patch changes the code so that ps->next_free skips the metadata area
when metadata are loaded in function read_exceptions.

The patch moves a piece of code from persistent_prepare_exception to a
separate function skip_metadata to avoid code duplication.

See the bug 995071 for a script that reproduces the bug and a simplified
exploit that shows security implications of the bug.

See the bug 974481 for original customer's report.

Signed-off-by: Mikulas Patocka <mpatocka@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>