nfs: Simplify the NFSv4 OPEN compound
Patchwork-id: 65878
O-Subject: [RHEL6.5 PATCH 09/12] NFSv4: Simplify the NFSv4 OPEN compound
Bugzilla: 998697
RH-Acked-by: Jeff Layton <jlayton@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Trond Myklebust <Trond.Myklebust@netapp.com>

Get rid of the post-op GETATTR on the directory in order to reduce
the amount of processing done on the server.

The cost is that if we later need to stat() the directory, then we
know that the ctime and mtime are likely to be invalid.

Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Upstream Id: 90ff0c548d1220d31f80e498b587393895705e6c
Red Hat Bz: 998697

Signed-off-by: Rafael Aquini <aquini@redhat.com>