igb: fix vlan filtering in promisc mode when not in VT mode
Patchwork-id: 66019
O-Subject: [RHEL6.5 PATCH] igb: fix vlan filtering in promisc mode when not in VT mode
Bugzilla: 996277
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>
RH-Acked-by: Nikolay Aleksandrov <nikolay@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Veaceslav Falico <vfalico@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Bugzilla:
https://bugzilla.redhat.com/show_bug.cgi?id=996277

Description:
This patch fixes a VT mode check to make sure VLAN filters are disabled when
in promisc mode and VT is not enabled.

The problem with the previous check was that:
E1000_MRQC_ENABLE_VMDQ is defined as 0x00000003

but when not in VT mode:
mrqc |= E1000_MRQC_ENABLE_RSS_4Q (0x00000002)

So the above check will trigger regardless if VT mode is being used or not.

Upstream Status:
http://git.kernel.org/?p=linux/kernel/git/davem/net-next.git;a=commitdiff;h=7e44892c1b6bb499cb2f6d5c0f4afcc077a26074

Brew:
https://brewweb.devel.redhat.com/taskinfo?taskID=6209896

Test:
Tested by jhsiao and myself.

Signed-off-by: Rafael Aquini <aquini@redhat.com>