iwlwifi: mvm: correctly configure MCAST in AP mode
Patchwork-id: 66276
O-Subject: [RHEL6.5 PATCH v2 34/48] iwlwifi: mvm: correctly configure MCAST in AP mode
Bugzilla: 992969
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

From: Emmanuel Grumbach <emmanuel.grumbach@intel.com>

BZ#992969
Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=6247982
Testing: Smoke tests by myself

commit 86a91ec757338edbce51de5dabd7afb0366f485c upstream.

The AP mode needs to use the MCAST fifo for the MCAST
frames sent after the DTIM. This fifo needs to be
configured with the same parameters as the VOICE FIFO.

A separate SCD queue is mapped to this fifo - the cab_queue
(cab stands for Content After Beacon). This queue isn't
connected to any station, but rather to the MAC context.
This queue should (and is already) be set as the MCAST
queue - this is part of the of MAC context command.

Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-by: Ilan Peer <ilan.peer@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>