configs: disable the 'ocrdma' driver
Patchwork-id: 66541
O-Subject: [RHEL6.5 PATCH] redhat/configs: disable the 'ocrdma' driver
Bugzilla: 1006303
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1006303

In bug 827601 we added the 'ocrdma' driver to RHEL 6.5. It turns out the
driver is not yet in a good shape upstream to be enabled in RHEL 6.5 GA.
We should disable it in the kernel configuration.

Quoting Doug Ledford
(https://bugzilla.redhat.com/show_bug.cgi?id=827601#c27):

After conversing with Emulex about the current state of this driver, we
know that there are additional patches on top of what we have already
integrated required before the driver is ready for use.  We also know
that there is an important set of patches needed upstream regarding GID
creation at the link layer level for IBoE and that both ocrdma and mlx4
will need to be updated to the new method in unison.  So, since this
driver is not yet complete, and since upstream has not yet reviewed and
accepted the additional patches needed to make it complete, I'm kicking
this bug back to ASSIGNED and it needs to be dropped from the kernel
errata and then moved out to rhel6.6 to be a placeholder for picking up
the remaining patches then.  In addition, I might recommend that we
disable the ocrdma driver build in rhel6.5 and re-enable the build when
we integrate the additional patches for rhel6.6

Testing:
I checked that ocrdma.ko is not built anymore.

Signed-off-by: Rafael Aquini <aquini@redhat.com>