fuse: drop dentry on failed revalidate
Patchwork-id: 66994
O-Subject: [RHEL6.5 PATCH 3/3] fuse: drop dentry on failed revalidate
Bugzilla: 924014
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

This is intended to be a backport of the following upstream commit,
but is significantly modified. The upstream patch uses a newly
introduced vfs dcache helper to walk/shrink a dentry subtree, check
for mounts and d_drop the entry. Other upstream fs' that require
this behavior have been modified to use this mechanism (NFS, GFS2).

The new vfs mechanism is too invasive for rhel6. As a result,
update fuse to implement this behavior as previously implemented
by NFS and GFS2 (and as currently implemented in rhel6).

commit 46ea1562da792a94ee8391f7aed882eb7771e18c
Author: Anand Avati <avati@redhat.com>
Date:   Thu Sep 5 11:44:44 2013 +0200

    fuse: drop dentry on failed revalidate

    Drop a subtree when we find that it has moved or been delated.  This can be
    done as long as there are no submounts under this location.

    If the directory was moved and we come across the same directory in a
    future lookup it will be reconnected by d_materialise_unique().

    Signed-off-by: Anand Avati <avati@redhat.com>
    Signed-off-by: Miklos Szeredi <mszeredi@suse.cz>
    Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>

Signed-off-by: Brian Foster <bfoster@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>