nvme: nvme_submit_bio_queue fixup check for DISCARD
Patchwork-id: 67115
O-Subject: [RHEL6.5 PATCH BZ 1006051 20/19] NVMe: nvme_submit_bio_queue fixup check for DISCARD
Bugzilla: 1006051
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Keith Busch <Keith.Busch@intel.com>

This patch is RHEL6 only, bug was discovered by Keith Busch during
testing. Here are his notes from BZ 1006051

The mke2fs tries to do a discard on the entirety of the device, and since the check
for REQ_DISCARD was failing, we were trying to map the bio's io vector for it's bi_size
as if it were a r/w command, which in this case bi_size was the entirety of the NVMe
namespace capacity, and so bad things happen. :)

Anyway, if my device doesn't report that it is capable of discard (or I apply the fix
above), everything works fine.

Signed-off-by: Rafael Aquini <aquini@redhat.com>