xen: mask cpu feature avx
Patchwork-id: 67303
O-Subject: [RHEL6.5 PATCH] xen: mask cpu feature avx
Bugzilla: 1006549
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Igor Mammedov <imammedo@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1006549
Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=6275797

We didn't have a problem with avx on rhel6 xen guests until
now, because that feature wasn't used by rhel6 until now.
6.5 commit 479f066d8ae2e changes that, so we need to mask off
the feature, otherwise we'll fail to install as the 2nd ramdisk
anaconda uses contains the xor module and attempts to load it.
We actually already mask this feature on the host-side, but the
patch that does that masking will only be in 5.10, so we need
to protect the guest with a guest-side patch as well. This
patch does the guest-side masking for both pv and fv guests.
Host-side, for AMD hosts, we've been masking since 5.7, so I
tested the fv part of this patch on an Intel machine with AVX.

Not upstream material. Tested by me.

Signed-off-by: Andrew Jones <drjones@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>