xfs: switch stacks for bmap btree modifications
Patchwork-id: 67325
O-Subject: [RHEL 6.5 PATCH] xfs: switch stacks for bmap btree modifications
Bugzilla: 918359
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>

From: Dave Chinner <dchinner@redhat.com>

RH BZ: 918359
Upstream commit: N/A

When backporting the upstrream stack switch code in the allocation
path, the changes were manually backported as the upstream code has
diverged significantly from the RHEL6 codebase. The upstream code
had been factored that made the boundaries for the allocationstack
switch obvious, but the RHEL6 code was still the old mess.

As a result, a part of the stack switch code was missed - the bmap
btree manipulation can cause further allocation and freeing of
blocks, and that can cause deeper stacks to occur and overrun.

Because the code is so different to upstream and there's no
equivalent upstream code commit, this is a RHEL6 only patch. it
takes part of the upstream factoring patches - the bit necessary to
isolate the entire string of allocations - and then moves the stack
switch to the caller of the newly factored function. This will push
all the allocations off to a separate stack and should fix the
problem being seen.

Signed-off-by: Dave Chinner <dchinner@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>