bonding: fix bond_arp_rcv setting and arp validate desync state
Patchwork-id: 67518
O-Subject: [RHEL6.5 PATCH BZ 1003697 3/3] bonding: fix bond_arp_rcv setting and arp validate desync state
Bugzilla: 1003697
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Veaceslav Falico <vfalico@redhat.com>

commit 5bb9e0b50d2188d8fac481742d9f801436e2c5ab
Author: nikolay@redhat.com <nikolay@redhat.com>
Date:   Sat Sep 7 00:00:26 2013 +0200

    bonding: fix bond_arp_rcv setting and arp validate desync state
    We make bond_arp_rcv global so it can be used in bond_sysfs if the bond
    interface is up and arp_interval is being changed to a positive value
    and cleared otherwise as per Jay's suggestion.
    This also fixes a problem where bond_arp_rcv was set even though
    arp_validate was disabled while the bond was up by unsetting recv_probe
    in bond_store_arp_validate and respectively setting it if enabled.

    Signed-off-by: Nikolay Aleksandrov <nikolay@redhat.com>
    Signed-off-by: Marcelo Ricardo Leitner <mleitner@redhat.com>
    Acked-by: Veaceslav Falico <vfalico@redhat.com>
    Signed-off-by: David S. Miller <davem@davemloft.net>

Note: This patch is based on the above commit but it's heavily re-worked
to fit RHEL6 such as the use of packet type hooks and different ways of
bringing up/down the bonding device, but essentially the functionality of
the upstream patch and this one is the same. Now the only way we can
have the hook registered is with arp_validate and arp_interval both set
so in bond_close we need to unregister it only in such case.

Signed-off-by: Rafael Aquini <aquini@redhat.com>