be2net: Clear any capability flags that driver is not interested in
Patchwork-id: 67526
O-Subject: [RHEL6.5 PATCH] be2net: Clear any capability flags that driver is not interested in.
Bugzilla: 998856
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Myron Stowe <mstowe@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

BZ: #998856 - be2net: Traffic stalls on Skyhawk with latest version of firmware

Description:
It is possible for some versions of firmware to advertise capabilities
that driver is not ready to handle. This may lead to controller stall.
Since the driver is interested only in subset of flags, clearing the rest.

Upstream commit:
3da988c be2net: Clear any capability flags that driver is not interested in.

Signed-off-by: Ivan Vecera <ivecera@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>