cma: Check for GID on listening device first
Patchwork-id: 67557
O-Subject: [Patch RHEL6 2/2] IB/cma: Check for GID on listening device first
Bugzilla: 920306
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

As a simple optimization that should speed up the vast majority of
connect attemps on IB devices, when we are searching for the GID of an
incoming connection in the cached GID lists of devices, search the
device that received the incoming connection request first.  If we don't
find it there, then move on to other devices.

Signed-off-by: Doug Ledford <dledford@redhat.com>
(cherry picked from commit 1986cf8f1a8edac566af563026bf4475b7ed1c30)
Signed-off-by: Doug Ledford <dledford@redhat.com>

Signed-off-by: Rafael Aquini <aquini@redhat.com>