vfs: allow umount to handle mountpoints without revalidating them
Patchwork-id: 67608
O-Subject: [RHEL-6.5 PATCH 1/5] vfs: allow umount to handle mountpoints without revalidating them
Bugzilla: 999708
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Jeff Layton <jlayton@redhat.com>

Bugzilla bug: https://bugzilla.redhat.com/show_bug.cgi?id=999708
Upstream commit: 8033426e6

For reference (for the comment below) see bugzilla bug:
https://bugzilla.redhat.com/show_bug.cgi?id=980172

Christopher reported a regression where he was unable to unmount a NFS
filesystem where the root had gone stale. The problem is that
d_revalidate handles the root of the filesystem differently from other
dentries, but d_weak_revalidate does not. We could simply fix this by
making d_weak_revalidate return success on IS_ROOT dentries, but there
are cases where we do want to revalidate the root of the fs.

A umount is really a special case. We generally aren't interested in
anything but the dentry and vfsmount that's attached at that point. If
the inode turns out to be stale we just don't care since the intent is
to stop using it anyway.

Try to handle this situation better by treating umount as a special
case in the lookup code. Have it resolve the parent using normal
means, and then do a lookup of the final dentry without revalidating
it. In most cases, the final lookup will come out of the dcache, but
the case where there's a trailing symlink or !LAST_NORM entry on the
end complicates things a bit.

Cc: Neil Brown <neilb@suse.de>
Reported-by: Christopher T Vogan <cvogan@us.ibm.com>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
Signed-off-by: Rafael Aquini <aquini@redhat.com>