gfs2: Fix race in iteration of glocks for unfreeze/umount
Patchwork-id: 67622
O-Subject: [RHEL6.5 PATCH][bz 999909] GFS2: Fix race in iteration of glocks for unfreeze/umount
Bugzilla: 999909
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

We need to check the glock ref counter in a race free way
in order to ensure that the gfs2_glock_hold() call will
succeed. The easiest way to do that is to simply take the
reference count early in the common code of examine_bucket,
skipping any glocks with zero ref count.

That means that the examiner functions all need to put their
reference on the glock once they've performed their function.

This is a RHEL6 port of the upstream patch here:
https://git.kernel.org/cgit/linux/kernel/git/steve/gfs2-3.0-nmw.git/commit/?id=7286b31eaba6404fa92f68d04626da1f395b3916

Resolves: rhbz#999909

Signed-off-by: Abhi Das <adas@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>