cnic: Fix crash in, cnic_bnx2x_service_kcq()
Patchwork-id: 67715
O-Subject: [RHEL6.5 e-stor PATCH 2/1] cnic: Fix crash in, cnic_bnx2x_service_kcq()
Bugzilla: 1004554
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Patch provide by Broadcom.

From upstream commit 48a30569732be782119960a47ca05417e7e07dab

changed the code to use the bnx2x macro NO_FCOE() to determine if FCoE
is supported or not.  There is another place in cnic that is still using
the old method to determine if FCoE is supported or not.  The 2 methods
may not yield the same result after the network interface is brought down
and up.  This will cause the crash as cnic_bnx2x_service_kcq() will access
the uninitialized cp->kcq2.

The fix is to consistently use the same macro CNIC_SUPPORTS_FCOE() which
uses the bnx2x NO_FCOE() macro.  As a follow-up, we can clean up the code
to remove the old method as it is no longer needed.

Signed-off-by: Michael Chan <mchan@broadcom.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Rafael Aquini <aquini@redhat.com>