cache: don't fill the cache with junk
Patchwork-id: 67764
O-Subject: [Patch RHEL6 3/3] IB/cache: don't fill the cache with junk
Bugzilla: 920306
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

We keep a cache of the GIDs and PKeys on an IB device, but when we get
the props for the card, the table length returned is the overall table
size and not related to how many valid entries there are in the table.
As a result, we end up with things like 128 entries for GIDs that are 3
valid GIDs and 125 empty GIDs.  Then when we call find_cache_gid, we
search through all 128 GID cache entries using memcmp.  This is
needlessly expensive.  So when we update the cache, check each item we
get from the card to see if it's valid and only save it into the cache
if it is.  We make sure to preserve the index from the card's table with
each cache item so the correct index in the card's table is returned
on any lookup that requests the index.

I have performance numbers on this change, but they aren't really
conclusive.  I made this change after the previous two patches, and
while conceptually it is obvious that search 3 or 4 GIDs is better than
searching through 128 empty GIDs using memcmp, the fact of the matter is
that we normally find our result quickly and so the benefit of this
change is not seen, at least not by using the cmtime application.  In
addition, the immediately previous patch optimized the connect routine's
selection of device to search first, again hiding the benefit of this
change.  It would require a very complex setup with lots of valid GIDs
and connect attempts that were accessing GIDs late in the table to
demonstrate the benefit of this patch clearly, and the cmtime utilitity
from librdmacm does not do that.

Signed-off-by: Doug Ledford <dledford@redhat.com>
(cherry picked from commit 4c88e00f5211787a98fa980a4d42c5c6374ab868)
Signed-off-by: Doug Ledford <dledford@redhat.com>

Signed-off-by: Rafael Aquini <aquini@redhat.com>