Add isync to copy_and_flush
Patchwork-id: 68339
O-Subject: [PATCH RHEL6.5 BZ1014475] powerpc: Add isync to copy_and_flush
Bugzilla: 1014475
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Nikolay Aleksandrov <nikolay@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1014475
Description:
------------
In __after_prom_start we copy the kernel down to zero in two calls to
copy_and_flush.  After the first call (copy from 0 to copy_to_here:)
we jump to the newly copied code soon after.

Unfortunately there's no isync between the copy of this code and the
jump to it.  Hence it's possible that stale instructions could still be
in the icache or pipeline before we branch to it.

We've seen this on real machines and it's results in no console output
after:
  calling quiesce...
  returning from prom_init

The below adds an isync to ensure that the copy and flushing has
completed before any branching to the new instructions occurs.

Signed-off-by: Michael Neuling <mikey@neuling.org>
CC: <stable@vger.kernel.org>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>

RHEL Version Found:
-------------------
RHEL 6.5

Brew:
-----
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=6359931

Upstream:
---------
http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=29ce3c5073057991217916abc25628e906911757

Test Status:
------------
Nishanth Aravamudan helped me test this.

---------------------------------------------------------------
Steve Best
IBM on-site partner

Signed-off-by: Rafael Aquini <aquini@redhat.com>