nfs: Send atime and mtime as a 64bit value
Patchwork-id: 68350
O-Subject: [RHEL6.6 PATCH] [BZ#996548] nfs: Send atime and mtime as a 64bit value
Bugzilla: 996548
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Bryan Schumaker <bjschuma@netapp.com>

RFC 3530 says that the seconds value of a nfstime4 structure is a 64bit
value, but we are instead sending a 32-bit 0 and then a 32bit conversion
of the 64bit Linux value.  This means that if we try to set atime to a
value before the epoch (touch -t 196001010101) the client will only send
part of the new value due to lost precision.

Signed-off-by: Bryan Schumaker <bjschuma@netapp.com>
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Upstream commit: 042ad0b398ea4e937e38324dd096bdf9d2c495f7

Signed-off-by: Rafael Aquini <aquini@redhat.com>