free bios when failing blk_execute_rq_nowait calls
Patchwork-id: 68366
O-Subject: [rhel6.5 patch] block: free bios when failing blk_execute_rq_nowait calls
Bugzilla: 1009312
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>

Mike Christie ran into the WARN_ON(req->bio != NULL) check in
__blk_put_request while doing some multipath and disk pull testing.  The
problem was introduced by the following commit in the 6.3 development
cycle:

commit 7738415f8e73a23d52243506ead5ea08bd958210
Author: Jeff Moyer <jmoyer@redhat.com>
Date:   Fri Feb 10 21:44:35 2012 -0500

    [block] fix crash in scsi_dispatch_cmd()

In that patch, logic was added to bail out early if the request queue
is marked dead:

@@ -50,6 +50,13 @@ void blk_execute_rq_nowait(struct request_queue *q, struct ge
 {
        int where = at_head ? ELEVATOR_INSERT_FRONT : ELEVATOR_INSERT_BACK;

+       if (unlikely(test_bit(QUEUE_FLAG_DEAD, &q->queue_flags))) {
+               rq->errors = -ENXIO;
+               if (rq->end_io)
+                       rq->end_io(rq, rq->errors);
+               return;
+       }
+

Unfortunately, that does not clean up the bios associated with the
struct request (and hence the warning).  Mike Christie provided a fix
upstream, and this is a backport of that fix:

commit 7652113c2f508b1c8176640dcd034730fe79bc48
Author: Mike Christie <michaelc@cs.wisc.edu>
Date:   Wed Sep 18 08:33:55 2013 -0600

    If the queue is dying then we only call the rq->end_io callout.
    This leaves bios setup on the request, because the caller assumes when
    the blk_execute_rq_nowait/blk_execute_rq call has completed that
    the rq->bios have been cleaned up.

    This patch has blk_execute_rq_nowait use __blk_end_request_all
    to free bios and also call rq->end_io.

    Signed-off-by: Mike Christie <michaelc@cs.wisc.edu>
    Signed-off-by: Jens Axboe <axboe@kernel.dk>

I ran a kernel with this patch through brew[1], and Mike has agreed to
test the patch (I've only boot tested it).  Given that this code is used
for testing multipath links (which happens periodically, on a timer),
it's a good idea to get this fixed for 6.5.

There is a slight difference from the upstream patch due to a global
rename of blk_queue_dead to blk_queue_dying by Bart Van Assche in commit
3f3299d.

This addresses bug 1009312.

[1] http://brewweb.devel.redhat.com/brew/taskinfo?taskID=6367635

Signed-off-by: Rafael Aquini <aquini@redhat.com>