libfcoe: Make fcoe_sysfs optional / fix fnic NULL exception
Patchwork-id: 68523
O-Subject: [RHEL6.5 PATCH] libfcoe: Make fcoe_sysfs optional / fix fnic NULL exception
Bugzilla: 1014864
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

fnic doesn't use any of the create/destroy/enable/disable interfaces
either from the (legacy) module paramaters or the (new) fcoe_sysfs
interfaces. When fcoe_sysfs was introduced fnic wasn't changed since
it wasn't using the interfaces. libfcoe incorrectly assumed that that
all of its users were using fcoe_sysfs and when adding and deleting
FCFs would assume the existance of a fcoe_ctlr_device. fnic was not
allocating this structure because it doesn't care about the standard
user interfaces (fnic starts on link only). If/When libfcoe tried to use
the fcoe_ctlr_device's lock for the first time a NULL pointer exception
would be triggered.

Since fnic doesn't care about sysfs or user interfaces, the solution
is to drop libfcoe's assumption that all drivers are using fcoe_sysfs.

This patch accomplishes this by changing some of the structure
relationships.

We need a way to determine when a LLD is using fcoe_sysfs or not and
we can do that by checking for the existance of the fcoe_ctlr_device.
Prior to this patch, it was assumed that the fcoe_ctlr structure was
allocated with the fcoe_ctlr_device and immediately followed it in
memory. To reach the fcoe_ctlr_device we would simply go back in memory
from the fcoe_ctlr to get the fcoe_ctlr_device.

Since fnic doesn't allocate the fcoe_ctlr_device, we cannot keep that
assumption. This patch adds a pointer from the fcoe_ctlr to the
fcoe_ctlr_device. For bnx2fc and fcoe we will continue to allocate the
two structures together, but then we'll set the ctlr->cdev pointer
to point at the fcoe_ctlr_device. fnic will not change and will continue
to allocate the fcoe_ctlr itself, and ctlr->cdev will remain NULL.

When libfcoe adds fcoe_fcf's to the fcoe_ctlr it will check if ctlr->cdev
is set and only if so will it continue to interact with fcoe_sysfs.

Cc: Hiral Patel <hiralpat@cisco.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>
Acked-by: Neil Horman <nhorman@tuxdriver.com>

BZ: 1014864

Upstream: In fcoe patchwork pending integration to the fcoe tree:
http://patchwork.open-fcoe.org/patch/104/

Brew: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=6371432

Test: Tested successfully by partner, who reported the issue
Notes: This is proposed as a 6.5 Blocker.  its not yet approved but I expect it
will be given that it causes fnic to panic the box on probe.  Note also that
this patch contains a slight change from the upstream version as Tomas Henzl
just noted a missing mutex_unlock in the upstream version, I've sent a
corrective patch upstream for it, and expect it will be merged with the
preceding patch

Signed-off-by: Rafael Aquini <aquini@redhat.com>