Fix skipping recovery for read-only arrays
Patchwork-id: 68540
O-Subject: [RHEL6.5 PATCH 1/1] md: Fix skipping recovery for read-only arrays.
Bugzilla: 1014102
RH-Acked-by: Doug Ledford <dledford@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Lukasz Dorau <lukasz.dorau@intel.com>

Since:
        commit 7ceb17e87bde79d285a8b988cfed9eaeebe60b86
        md: Allow devices to be re-added to a read-only array.

spares are activated on a read-only array. In case of raid1 and raid10
personalities it causes that not-in-sync devices are marked in-sync
without checking if recovery has been finished.

If a read-only array is degraded and one of its devices is not in-sync
(because the array has been only partially recovered) recovery will be skipped.

This patch adds checking if recovery has been finished before marking a device
in-sync for raid1 and raid10 personalities. In case of raid5 personality
such condition is already present (at raid5.c:6029).

Bug was introduced in 3.10 and causes data corruption.

Cc: stable@vger.kernel.org
Signed-off-by: Pawel Baldysiak <pawel.baldysiak@intel.com>
Signed-off-by: Lukasz Dorau <lukasz.dorau@intel.com>
Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Posted upstream:
http://marc.info/?l=linux-raid&m=138115595902520&w=4

Signed-off-by: Rafael Aquini <aquini@redhat.com>