cifs: only set ops for inodes in I_NEW state
Patchwork-id: 69483
O-Subject: [RHEL6.6 PATCH 1/3][bz1017883] cifs: only set ops for inodes in I_NEW state
Bugzilla: 1017883
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Jeff Layton <jlayton@redhat.com>

It's generally not safe to reset the inode ops once they've been set. In
the case where the inode was originally thought to be a directory and
then later found to be a DFS referral, this can lead to an oops when we
try to trigger an inode op on it after changing the ops to the blank
referral operations.

Cc: <stable@vger.kernel.org>
Reported-and-Tested-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <smfrench@gmail.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
(cherry picked from commit 62106e96279f66c52e2123782f9420af9dbe8cbe)

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>