async: Bump up the MAX_THREADS count for the async subsystem
Patchwork-id: 69493
O-Subject: [RHEL6.5 PATCH] async: Bump up the MAX_THREADS count for the async subsystem
Bugzilla: 1010666
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: WANG Chao <chaowang@redhat.com>

BZ 1010666 reported a 3-way deadlock, in which the async thread was waiting on a
workqueue flush, a workqueue thread was waiting on an error handler thread, and
the error handler thread was working on async thread compeltion.

The problem was something of a perfect storm, created by the facts that:

1) kdump runs in single cpu mode preventing forward progress enabled by multiple
parallel execution

2) my recent commit b727b50, which calls schedule inside the async_task
function, making it possible for multiple threads to schedule while its running

3) the sas high availability api which allows blocking on workqueue threads that
in turn block on async threads

4) Our workqueue implementation, which doesn't have the upstream changes to
allow flushing from within a workqueue or async task (too much risk to
backport).

So we need to find a way to break the dependency chain.  I'm not comfortable
reverting b727b50, as that would cause a regression of bz 817955, and even if we
removed it, its still posible for that thread to get pre-empted, which could
cause this bug again.  We can't run kdump with multiple cpus for any number of
reasons, and we can't backport the upstream workqueue changes.

Fortunately we have an out here.  The scsi error handler uses a private aysnc
domain to complete, so as long as there is one async thread free to do work, we
can make forward progress.  This patch bumps up the MAX_THREADS value for async
to at least 5 threads, so one will be available if the others are blocked.  5 is
something of an arbitrary value, but we just need a small pool of threads to
allow forward work to progress.

This RHEL6 specific patch implements the above solution.

BZ: 1010666
Testing: Tested by the reporter with success
Brew: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=6397345
Upstream: NA, RHEL specific patch
Signed-off-by: Rafael Aquini <aquini@redhat.com>